+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 0.0.* (2020-11-05)
+ * CRUD for MenuItems models
  * Assets, menu item form
  * CRUD for Menu models
  * Added base module, migrations and models